DPL Analysis: move Configurable to a separate header
* adapt LiteralNode to accept Configurable
