+++ Framework/Core/include/Framework/AnalysisHelpers.h
@@ -428,28 +428,6 @@ struct Service {
   }
 };
 
-/// This helper allows you to create a configurable option associated to a task.
-/// Internally it will be bound to a ConfigParamSpec.
-template <typename T>
-struct Configurable {
-  Configurable(std::string const& name, T defaultValue, std::string const& help)
-    : name(name), value(defaultValue), help(help)
-  {
-  }
-  using type = T;
-  std::string name;
-  T value;
-  std::string help;
-  operator T()
-  {
-    return value;
-  }
-  T const* operator->() const
-  {
-    return &value;
-  }
-};
-
 template <typename T>
 struct Partition {
   Partition(expressions::Node&& node)
